Moved grid map creation to hook
Updated tests

Ref #264
